backend/requirements.txt
mss
jax[cuda12]
pillow